+++ sdk/nodejs/ram/index.ts
@@ -2,4 +2,6 @@
 // *** Do not edit by hand unless you're certain you know what you are doing! ***
 
 // Export members:
+export * from "./principalAssociation";
+export * from "./resourceAssociation";
 export * from "./resourceShare";