+++ sdk/nodejs/ses/receiptRule.ts
@@ -18,12 +18,14 @@ import * as utilities from "../utilities";
  *     addHeaderActions: [{
  *         headerName: "Custom-Header",
  *         headerValue: "Added by SES",
+ *         position: 1,
  *     }],
  *     enabled: true,
  *     recipients: ["karen@example.com"],
  *     ruleSetName: "default-rule-set",
  *     s3Actions: [{
  *         bucketName: "emails",
+ *         position: 2,
  *     }],
  *     scanEnabled: true,
  * });