+++ sdk/python/pulumi_aws/docdb/__init__.py
@@ -4,6 +4,7 @@
 
 # Export this package's modules as members:
 from .cluster import *
+from .cluster_instance import *
 from .cluster_parameter_group import *
 from .cluster_snapshot import *
 from .subnet_group import *