+++ sdk/python/pulumi_aws/ram/__init__.py
@@ -3,4 +3,6 @@
 # *** Do not edit by hand unless you're certain you know what you are doing! ***
 
 # Export this package's modules as members:
+from .principal_association import *
+from .resource_association import *
 from .resource_share import *