+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.16.9 (Unreleased)
 
+### Improvements
+
+- Updated to the v1.59.0 version of the AWS Terraform Provider.
+
 ## 0.16.8 (Released February 11th, 2019)